feat: Added unit testing support. Moved the client definition to client.ts for easier mocking. Resolves #619
