+++ .gitignore
@@ -8,3 +8,4 @@ database/toDeleteList.json
 database/voteCache.json
 database/webhookCache.json
 dist
+report
