+++ package.json
@@ -1,35 +1,41 @@
 {
-	"name": "paw-and-paper",
-	"version": "3.1.2",
-	"description": "A Discord Bot for roleplay communities, featuring a community-based interactive RPG and powerful tools to roleplay",
-	"main": "index.ts",
-	"scripts": {
-		"start": "git pull origin stable && npm update && tsc -p tsconfig.json && pm2 start dist/index.js --name paw --time",
-		"reload": "git pull origin stable && npm update && tsc -p tsconfig.json && pm2 reload paw",
-		"stop": "pm2 stop paw",
-		"logs": "pm2 logs --timestamp --lines 0 paw",
-		"test": "ts-node src/index.ts"
-	},
-	"author": "Maksi",
-	"license": "ISC",
-	"dependencies": {
-		"@octokit/rest": "^19.0.5",
-		"@top-gg/sdk": "^3.1.3",
-		"bfd-api-redux": "^1.2.2",
-		"crystalid": "^1.0.1",
-		"discord.js": "^14.6.0",
-		"express": "^4.18.2",
-		"express-rate-limit": "^6.6.0",
-		"fs": "^0.0.1-security",
-		"fuse.js": "^6.6.2",
-		"hoatzin": "^1.3.3",
-		"topgg-autoposter": "^2.0.1"
-	},
-	"devDependencies": {
-		"@types/express": "^4.17.14",
-		"@typescript-eslint/eslint-plugin": "^5.41.0",
-		"@typescript-eslint/parser": "^5.41.0",
-		"eslint": "^8.26.0",
-		"ts-node": "^10.9.1"
-	}
+  "name": "paw-and-paper",
+  "version": "3.1.2",
+  "description": "A Discord Bot for roleplay communities, featuring a community-based interactive RPG and powerful tools to roleplay",
+  "main": "index.ts",
+  "scripts": {
+    "start": "git pull origin stable && npm update && tsc -p tsconfig.json && pm2 start dist/index.js --name paw --time",
+    "reload": "git pull origin stable && npm update && tsc -p tsconfig.json && pm2 reload paw",
+    "stop": "pm2 stop paw",
+    "logs": "pm2 logs --timestamp --lines 0 paw",
+    "test": "jest & node testReportMessage.mjs",
+    "startdev": "ts-node src/index.ts"
+  },
+  "author": "Maksi",
+  "license": "ISC",
+  "dependencies": {
+    "@octokit/rest": "^19.0.5",
+    "@top-gg/sdk": "^3.1.3",
+    "bfd-api-redux": "^1.2.2",
+    "crystalid": "^1.0.1",
+    "discord.js": "^14.6.0",
+    "express": "^4.18.2",
+    "express-rate-limit": "^6.6.0",
+    "fs": "^0.0.1-security",
+    "fuse.js": "^6.6.2",
+    "hoatzin": "^1.3.3",
+    "topgg-autoposter": "^2.0.1"
+  },
+  "devDependencies": {
+    "@johanblumenberg/ts-mockito": "^1.0.34",
+    "@types/express": "^4.17.14",
+    "@types/jest": "^29.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
+    "eslint": "^8.26.0",
+    "jest": "^29.2.2",
+    "jest-stare": "^2.4.1",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1"
+  }
 }
