+++ src/commands/quid_customization/profile.ts
@@ -7,7 +7,8 @@ import { hasCooldown, checkResting } from '../../utils/checkValidity';
 import { getMapData } from '../../utils/helperFunctions';
 import { disableCommandComponent } from '../../utils/componentDisabling';
 import { hasPermission, missingPermissions } from '../../utils/permissionHandler';
-import { client, commonPlantsInfo, materialsInfo, rarePlantsInfo, specialPlantsInfo, speciesInfo, uncommonPlantsInfo } from '../..';
+import { client } from '../../client';
+import { commonPlantsInfo, materialsInfo, rarePlantsInfo, specialPlantsInfo, speciesInfo, uncommonPlantsInfo } from '../..';
 import { SlashCommand } from '../../typings/handle';
 import { CurrentRegionType, RankType, UserData } from '../../typings/data/user';
 import { constructCustomId, constructSelectOptions, deconstructCustomId, deconstructSelectOptions } from '../../utils/customId';
