+++ src/commands_guild/957078956820033606/restart.ts
@@ -1,6 +1,6 @@
 import { PermissionFlagsBits, SlashCommandBuilder, User } from 'discord.js';
 import { respond } from '../../utils/helperFunctions';
-import { client } from '../..';
+import { client } from '../../client';
 import { SlashCommand } from '../../typings/handle';
 
 export const command: SlashCommand = {
