+++ src/events/interactionCreate.ts
@@ -11,7 +11,8 @@ import { sendErrorMessage } from '../utils/helperFunctions';
 import { generateId } from 'crystalid';
 import { readFileSync, writeFileSync } from 'fs';
 import { missingPermissions } from '../utils/permissionHandler';
-import { client, handle } from '../index';
+import { client } from '../client';
+import { handle } from '../index';
 import { ErrorStacks } from '../typings/data/general';
 import { hasNameAndSpecies } from '../utils/checkUserState';
 import { sendReminder, stopReminder } from '../commands/gameplay_maintenance/water-tree';
